+++ app/src/androidTest/kotlin/dev/thiagosouto/marvelpoc/base/BaseRobot.kt
@@ -9,20 +9,21 @@ import androidx.compose.ui.test.hasTestTag
 import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.compose.ui.test.onNodeWithTag
 
-internal abstract class BaseRobot(protected val rule: ComposeTestRule) {
+internal abstract class BaseRobot(protected val rule: ComposeTestRule) :
+    Retryable by RetryRobot(rule) {
 
     private fun ancestor(tag: String) = hasAnyAncestor(hasTestTag(tag))
 
     protected fun SemanticsNodeInteraction.waitUntilVisible() = apply {
-        rule.waitUntil(1000L) { isValid { this@waitUntilVisible.assertIsDisplayed() } }
+        rule.waitUntil(LONG_TIMEOUT) { isValid { this@waitUntilVisible.assertIsDisplayed() } }
     }
 
     protected fun SemanticsNodeInteraction.waitUntilDoesNotExist() = apply {
-        rule.waitUntil(1000L) { isValid { this@waitUntilDoesNotExist.assertDoesNotExist() } }
+        rule.waitUntil(LONG_TIMEOUT) { isValid { this@waitUntilDoesNotExist.assertDoesNotExist() } }
     }
 
     protected fun SemanticsNodeInteraction.waitUntilIsNotDisplayed() = apply {
-        rule.waitUntil(1000L) { isValid { this@waitUntilIsNotDisplayed.assertIsNotDisplayed() } }
+        rule.waitUntil(LONG_TIMEOUT) { isValid { this@waitUntilIsNotDisplayed.assertIsNotDisplayed() } }
     }
 
     private fun isValid(passedFunction: () -> Unit): Boolean =
@@ -35,60 +36,6 @@ internal abstract class BaseRobot(protected val rule: ComposeTestRule) {
             false
         }
 
-    protected fun retry(count: Int = 0, max: Int = 5, func: ComposeTestRule.() -> Unit) {
-        try {
-            rule.func()
-        } catch (e: ComposeTimeoutException) {
-            retryException(e, count, max, func)
-        } catch (e: AssertionError) {
-            retryException(e, count, max, func)
-        }
-    }
-
-    private fun retryException(
-        e: Throwable,
-        count: Int = 0,
-        max: Int = 5,
-        func: ComposeTestRule.() -> Unit
-    ) {
-        if (count == max) {
-            throw e
-        }
-        retry(count + 1, max, func)
-    }
-
-    protected fun retryWithDelay(
-        count: Int = 0,
-        max: Int = 5,
-        delay: Long = 50L,
-        func: ComposeTestRule.() -> Unit
-    ) {
-        try {
-            rule.func()
-        } catch (e: ComposeTimeoutException) {
-            retryException(delay, e, count + 1, max, func)
-        } catch (e: AssertionError) {
-            retryException(delay, e, count + 1, max, func)
-        }
-    }
-
-    private fun retryException(
-        delay: Long,
-        e: Throwable,
-        count: Int = 0,
-        max: Int = 5,
-        func: ComposeTestRule.() -> Unit
-    ) {
-        if (count == max) {
-            throw e
-        }
-
-        if (delay != 0L) {
-            Thread.sleep(delay)
-        }
-        retryWithDelay(count + 1, max, delay, func)
-    }
-
     fun applyComposable(func: ComposeTestRule.() -> Unit) = apply {
         rule.func()
     }
@@ -97,5 +44,9 @@ internal abstract class BaseRobot(protected val rule: ComposeTestRule) {
         rule
             .onNodeWithTag(tag)
             .assertDoesNotExist()
+    }
+
+    private companion object {
+        const val LONG_TIMEOUT = 1000L
     }
 }
